IsThisAValidTeamName/requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
Django==2.2.6
idna==2.8
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.0.2
multitasking==0.0.9
numpy==1.17.2
pandas==0.25.1
pandas-datareader==0.8.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
djangorestframework==3.10.3
sqlparse==0.3.0
urllib3==1.25.6
yfinance==0.1.45
praw==6.4.0
